requirements.txt
Django==5.1.8
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
PyJWT~=2.10.1
drf-spectacular==0.28.0
requests==2.32.3
pika==1.3.2
django-environ==0.12.0
environ==1.0
tenacity==9.0.0
pybreaker==1.2.0
django-cors-headers==4.7.0